+++ DemoAppSwiftUI/PinChannelHelpers.swift
@@ -150,15 +150,27 @@ struct DemoAppChatChannelNavigatableListItem<ChannelDestination: View>: View {
 
     public var body: some View {
         ZStack {
-            DemoAppChatChannelListItem(
-                channel: channel,
-                channelName: channelName,
-                injectedChannelInfo: injectedChannelInfo,
-                avatar: avatar,
-                onlineIndicatorShown: onlineIndicatorShown,
-                disabled: disabled,
-                onItemTap: onItemTap
-            )
+            if AppConfiguration.default.isChannelPinningFeatureEnabled {
+                DemoAppChatChannelListItem(
+                    channel: channel,
+                    channelName: channelName,
+                    injectedChannelInfo: injectedChannelInfo,
+                    avatar: avatar,
+                    onlineIndicatorShown: onlineIndicatorShown,
+                    disabled: disabled,
+                    onItemTap: onItemTap
+                )
+            } else {
+                ChatChannelListItem(
+                    channel: channel,
+                    channelName: channelName,
+                    injectedChannelInfo: injectedChannelInfo,
+                    avatar: avatar,
+                    onlineIndicatorShown: onlineIndicatorShown,
+                    disabled: disabled,
+                    onItemTap: onItemTap
+                )
+            }
 
             NavigationLink(
                 tag: channel.channelSelectionInfo,