+++ Sources/StreamChatSwiftUI/ChatChannelList/ChannelListConfig.swift
@@ -8,10 +8,12 @@ import Foundation
 public struct ChannelListConfig {
     public init(
         messageRelativeDateFormatEnabled: Bool = false,
-        showChannelListDividerOnLastItem: Bool = true
+        showChannelListDividerOnLastItem: Bool = true,
+        channelItemMutedStyle: ChannelItemMutedLayoutStyle = .default
     ) {
         self.messageRelativeDateFormatEnabled = messageRelativeDateFormatEnabled
         self.showChannelListDividerOnLastItem = showChannelListDividerOnLastItem
+        self.channelItemMutedStyle = channelItemMutedStyle
     }
 
     /// If true, the timestamp format depends on the time passed.
@@ -24,4 +26,7 @@ public struct ChannelListConfig {
     ///
     /// By default, all items in the channel list have a divider, including the last item.
     public var showChannelListDividerOnLastItem: Bool
+
+    /// The style for the channel item when it is muted.
+    public var channelItemMutedStyle: ChannelItemMutedLayoutStyle = .default
 }